Move `PhysFrame` type into new `frame` submodule
